fix: update outcome reconciliation after scheduler merge
